long type casting

+++ src/main/java/com/phonedirectory/app/config/JWTTokenUtil.java
@@ -18,9 +18,9 @@ import io.jsonwebtoken.SignatureAlgorithm;
 @Component
 public class JWTTokenUtil implements Serializable {
 	
-	private static final long serialVersionUID = -2550185165626007488L;
+	private static final long serialVersionUID = (long)(-2550185165626007488L);
 
-	public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60;
+	public static final long JWT_TOKEN_VALIDITY = (long)(5 * 60 * 60);
 
 	@Value("${jwt.secret}")
 	private String secret;
